Increase Module and Fix Some Infinitesimal Bugs
