+++ src/index.html
@@ -14,7 +14,7 @@
 		<header>
 			<div class="to-do__title flex-row border-bottom">
 				<h3>Today's To Do</h3>
-				<span class="material-symbols-outlined"> autorenew </span>
+				<span class="material-symbols-outlined clearAll"> autorenew </span>
 			</div>
 			<div class="to-do__add flex-row border-bottom">
 				<input
