+++ src/modules/addTask.js
@@ -1,5 +1,6 @@
 import renderToDoItems from './renderTasks.js';
 import toDoItemsArray from './toDoItemsArray.js';
+import init from './init.js';
 
 const addElBtn = document.querySelector('.to-do__addTask');
 const todoinputEl = document.querySelector('.to-do__input');
@@ -16,6 +17,7 @@ const addTask = () => {
       toDoItemsArray.push(newTodo);
       localStorage.setItem('toDoItemsArray', JSON.stringify(toDoItemsArray));
       renderToDoItems();
+      init();
       todoinputEl.value = '';
     } else {
       input.classList.add('inputError');
